+++ CHANGELOG.md
@@ -29,6 +29,8 @@ Changes can also be flagged with a GitHub label for tracking purposes. The URL o
 - Changed Fides.js banner title attributes for better SEO + a11y support [#6470](https://github.com/ethyca/fides/pull/6470)
 - Test Website Monitors now show the brand image and a "test monitor" tag [#6476](https://github.com/ethyca/fides/pull/6476)
 - Privacy Request runner will now also check for awaiting input tags as part of `poll_for_exited_privacy_request_tasks` [#6471](https://github.com/ethyca/fides/pull/6471)
+- Added async polling queue for API integrations [#6435](https://github.com/ethyca/fides/pull/6435)
+
 
 ## [2.68.0](https://github.com/ethyca/fides/compare/2.67.2...2.68.0)